lstv_be/requirements.txt
aiohttp==3.7.2
alive-progress==1.3.3
amqp==2.5.2
appnope==0.1.0
APScheduler==3.6.3
argh==0.26.2
asgiref==3.2.10
astroid==2.3.3
async-generator==1.10
async-timeout==3.0.1
attrs==19.3.0
Babel==2.8.0
backcall==0.1.0
bcrypt==3.1.7
beautifulsoup4==4.9.3
billiard==3.6.1.0
bleach==3.1.4
blessed==1.17.0
boto3==1.14.53
botocore==1.17.53
bugsnag==3.6.0
cached-property==1.5.1
cachetools==4.1.1
cattrs==1.0.0
celery==4.3.0
cement==3.0.4
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
click==6.7
click-completion==0.5.2
colorama==0.4.3
coreschema==0.0.4
crayons==0.4.0
cryptography==2.8
decorator==4.4.1
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.1
django-cors-headers==3.2.1
django-db-geventpool==3.2.0
django-elasticsearch-dsl==7.1.1
django-enum-choices==2.1.1
django-enumchoicefield==2.0.0
django-extensions==2.2.6
django-hashers-passlib==0.4
django-jsonfield==1.4.0
django-redis==4.11.0
django-rest-framework-condition==0.1.1
django-samesite-none==0.0.3
djangorestframework==3.11.0
dnspython==1.16.0
docker==4.1.0
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
drf-jwt==1.17.1
elasticsearch==7.5.1
elasticsearch-dsl==7.1.0
email-validator==1.1.1
entrypoints==0.3
eventlet==0.29.1
facebook-sdk==3.1.0
ffmpeg-python==0.2.0
Flask==1.1.1
flower==0.9.3
future==0.18.2
google-api-core==1.22.1
google-api-python-client==1.11.0
google-auth==1.21.0
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
greenlet==0.4.15
gunicorn==20.0.4
html2text==2020.1.16
httplib2==0.18.1
idna==2.7
importlib-metadata==1.5.0
importlib-resources==1.0.2
iniconfig==1.1.1
install==1.3.3
ipykernel==5.1.4
ipython==7.12.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isodate==0.6.0
isort==4.3.21
itsdangerous==0.24
itypes==1.2.0
jedi==0.16.0
Jinja2==2.11.1
jmespath==0.9.4
joblib==0.16.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.1.0
jupyter-core==4.6.1
jupyterlab-pygments==0.1.0
jwplatform==2.1.1
kombu==4.6.7
lazy-object-proxy==1.4.3
limits==1.5
mailchimp3==3.0.14
MarkupSafe==0.23
mccabe==0.6.1
mistune==2.0.0a2
monotonic==1.5
more-itertools==8.2.0
multidict==5.0.0
nbformat==5.0.4
nested-lookup==0.2.21
neterr==1.1.1
nltk==3.5
notebook==6.0.3
numpy==1.18.1
oauthlib==3.1.0
packaging==20.4
pandocfilters==1.4.2
paramiko==2.7.1
parso==0.6.0
passlib==1.7.2
pathspec==0.7.0
pathtools==0.1.2
pexpect==4.8.0
pi==0.1.2
pickleshare==0.7.5
Pillow==7.1.1
plaid-python==3.5.0
pluggy==0.13.1
powerline-status==2.7
prometheus-client==0.7.1
prompt-toolkit==3.0.3
protobuf==3.13.0
psycogreen==1.0.1
psycopg2==2.8.4
psycopg2-binary==2.8.4
ptyprocess==0.6.0
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
pycycle==0.0.8
Pygments==2.5.2
pyinstrument-cext==0.2.2
PyJWT==1.7.1
pylint==2.4.4
pylogbeat==1.0.2
PyNaCl==1.3.0
pyparsing==2.4.7
pyrsistent==0.15.7
pytest==6.1.2
python-dateutil==2.8.1
python-http-client==3.2.7
python-logstash-async==1.6.4
python-slugify==4.0.0
pytz==2019.3
PyYAML==5.3
pyzmq==18.1.1
qtconsole==4.6.0
redis==3.4.1
regex==2020.7.14
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.6
s3transfer==0.3.3
semantic-version==2.8.4
Send2Trash==1.5.0
sendgrid==6.6.0
shellingham==1.3.2
six==1.14.0
slackclient==2.9.3
soupsieve==2.1
sqlparse==0.3.0
starkbank-ecdsa==1.1.0
tabulate==0.8.6
tenacity==6.2.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.3
texttable==1.6.2
timezonefinder==4.2.0
toml==0.10.2
tornado==5.1.1
tqdm==4.48.2
traitlets==4.3.3
twilio==6.35.5
typed-ast==1.4.1
typing==3.7.4.3
typing-extensions==3.7.4.3
tzlocal==2.0.0
Unidecode==1.1.1
uritemplate==3.0.1
urllib3==1.24.3
validators==0.14.2
vine==1.3.0
watchdog==0.10.1
wcwidth==0.1.8
webencodings==0.5.1
WebOb==1.8.6
websocket-client==0.51.0
Werkzeug==1.0.0
wget==3.2
widgetsnbextension==3.5.1
wrapt==1.11.2
yarl==1.6.2
zerobouncesdk==0.0.6
zipp==2.1.0
